archive/old-deployment-scripts/extracted_code/app/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
PyJWT>=2.6.0
passlib>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
bcrypt>=4.0.1
alembic>=1.10.0
pytest>=7.3.1
pytest-timeout>=2.2.0
httpx>=0.24.0
psycopg2-binary>=2.9.5
sqlalchemy-utils>=0.40.0
websockets==10.4
uvicorn[standard]
python-jose>=3.3.0
requests>=2.31.0
email-validator>=2.0.0
